+++ internal/bind/bind.go
@@ -1,4 +1,4 @@
-package query
+package bind
 
 import (
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/allocator"
@@ -19,7 +19,7 @@ func (bindings Bindings) RewriteQuery(query string, args ...interface{}) (
 ) {
 	if len(bindings) == 0 {
 		var params []table.ParameterOption
-		params, err = ToYdb(args...)
+		params, err = Params(args...)
 		if err != nil {
 			return "", nil, xerrors.WithStackTrace(err)
 		}
@@ -36,7 +36,7 @@ func (bindings Bindings) RewriteQuery(query string, args ...interface{}) (
 		}
 	}
 
-	params, err := ToYdb(args...)
+	params, err := Params(args...)
 	if err != nil {
 		return "", nil, xerrors.WithStackTrace(err)
 	}
