+++ internal/bind/positional_args_test.go
@@ -1,4 +1,4 @@
-package query
+package bind
 
 import (
 	"testing"
@@ -10,10 +10,10 @@ import (
 	"github.com/ydb-platform/ydb-go-sdk/v3/table/types"
 )
 
-func TestNumericArgsBindRewriteQuery(t *testing.T) {
+func TestPositionalArgsBindRewriteQuery(t *testing.T) {
 	var (
 		now = time.Now()
-		b   = NumericArgsBind{}
+		b   = PositionalArgs{}
 	)
 	for _, tt := range []struct {
 		sql    string
@@ -23,64 +23,12 @@ func TestNumericArgsBindRewriteQuery(t *testing.T) {
 		err    error
 	}{
 		{
-			sql: `SELECT $123abc, $2`,
-			args: []interface{}{
-				100,
-			},
-			err: ErrInconsistentArgs,
-		},
-		{
-			sql: `SELECT $123abc, $1`,
-			args: []interface{}{
-				200,
-			},
-			yql: `-- origin query with numeric args replacement
-SELECT $123abc, $p0`,
-			params: []interface{}{
-				table.ValueParam("$p0", types.Int32Value(200)),
-			},
-		},
-		{
-			sql: `SELECT $1, $2`,
-			args: []interface{}{
-				table.ValueParam("$name1", types.Int32Value(100)),
-				table.ValueParam("$name2", types.Int32Value(200)),
-			},
-			yql: `-- origin query with numeric args replacement
-SELECT $name1, $name2`,
-			params: []interface{}{
-				table.ValueParam("$name1", types.Int32Value(100)),
-				table.ValueParam("$name2", types.Int32Value(200)),
-			},
-		},
-		{
-			sql: `SELECT $1, $2`,
-			args: []interface{}{
-				table.ValueParam("$namedArg", types.Int32Value(100)),
-				200,
-			},
-			yql: `-- origin query with numeric args replacement
-SELECT $namedArg, $p1`,
-			params: []interface{}{
-				table.ValueParam("$namedArg", types.Int32Value(100)),
-				table.ValueParam("$p1", types.Int32Value(200)),
-			},
-		},
-		{
-			sql: `SELECT $0, $1`,
+			sql: `SELECT ?, ?`,
 			args: []interface{}{
 				100,
 				200,
 			},
-			err: ErrUnexpectedNumericArgZero,
-		},
-		{
-			sql: `SELECT $1, $2`,
-			args: []interface{}{
-				100,
-				200,
-			},
-			yql: `-- origin query with numeric args replacement
+			yql: `-- origin query with positional args replacement
 SELECT $p0, $p1`,
 			params: []interface{}{
 				table.ValueParam("$p0", types.Int32Value(100)),
@@ -88,51 +36,50 @@ SELECT $p0, $p1`,
 			},
 		},
 		{
-			sql: `SELECT $1, $2`,
+			sql: `SELECT ?, ?`,
 			args: []interface{}{
 				100,
 			},
 			err: ErrInconsistentArgs,
 		},
 		{
-			sql: `SELECT $1, "$2"`,
+			sql: `SELECT ?, "?"`,
 			args: []interface{}{
 				100,
 			},
-			yql: `-- origin query with numeric args replacement
-SELECT $p0, "$2"`,
+			yql: `-- origin query with positional args replacement
+SELECT $p0, "?"`,
 			params: []interface{}{
 				table.ValueParam("$p0", types.Int32Value(100)),
 			},
 		},
 		{
-			sql: `SELECT $1, '$2'`,
+			sql: `SELECT ?, '?'`,
 			args: []interface{}{
 				100,
 			},
-			yql: `-- origin query with numeric args replacement
-SELECT $p0, '$2'`,
+			yql: `-- origin query with positional args replacement
+SELECT $p0, '?'`,
 			params: []interface{}{
 				table.ValueParam("$p0", types.Int32Value(100)),
 			},
 		},
 		{
-			sql: "SELECT $1, `$2`",
+			sql: "SELECT ?, `?`",
 			args: []interface{}{
 				100,
 			},
-			yql: "-- origin query with numeric args replacement\nSELECT $p0, `$2`",
+			yql: "-- origin query with positional args replacement\nSELECT $p0, `?`",
 			params: []interface{}{
 				table.ValueParam("$p0", types.Int32Value(100)),
 			},
 		},
 		{
-			sql:    "SELECT ?, $1, $p0",
-			params: []interface{}{},
-			err:    ErrInconsistentArgs,
+			sql: "SELECT ?, $1, $p0",
+			err: ErrInconsistentArgs,
 		},
 		{
-			sql: "SELECT $1, $2, $3",
+			sql: "SELECT ?, ?, ?",
 			args: []interface{}{
 				1,
 				"test",
@@ -142,7 +89,7 @@ SELECT $p0, '$2'`,
 					"test3",
 				},
 			},
-			yql: `-- origin query with numeric args replacement
+			yql: `-- origin query with positional args replacement
 SELECT $p0, $p1, $p2`,
 			params: []interface{}{
 				table.ValueParam("$p0", types.Int32Value(1)),
@@ -155,30 +102,7 @@ SELECT $p0, $p1, $p2`,
 			},
 		},
 		{
-			sql: "SELECT $1, $2, $3, $1, $2",
-			args: []interface{}{
-				1,
-				"test",
-				[]string{
-					"test1",
-					"test2",
-					"test3",
-				},
-			},
-			yql: `-- origin query with numeric args replacement
-SELECT $p0, $p1, $p2, $p0, $p1`,
-			params: []interface{}{
-				table.ValueParam("$p0", types.Int32Value(1)),
-				table.ValueParam("$p1", types.TextValue("test")),
-				table.ValueParam("$p2", types.ListValue(
-					types.TextValue("test1"),
-					types.TextValue("test2"),
-					types.TextValue("test3"),
-				)),
-			},
-		},
-		{
-			sql: "SELECT $1, $2, $3",
+			sql: "SELECT ?, ?, ?",
 			args: []interface{}{
 				types.Int32Value(1),
 				types.TextValue("test"),
@@ -188,7 +112,7 @@ SELECT $p0, $p1, $p2, $p0, $p1`,
 					types.TextValue("test3"),
 				),
 			},
-			yql: `-- origin query with numeric args replacement
+			yql: `-- origin query with positional args replacement
 SELECT $p0, $p1, $p2`,
 			params: []interface{}{
 				table.ValueParam("$p0", types.Int32Value(1)),
@@ -201,12 +125,12 @@ SELECT $p0, $p1, $p2`,
 			},
 		},
 		{
-			sql: "SELECT $1, a, b, c WHERE id = $1 AND date < $2 AND value IN ($3)",
+			sql: "SELECT a, b, c WHERE id = ? AND date < ? AND value IN (?)",
 			args: []interface{}{
 				1, now, []string{"3"},
 			},
-			yql: `-- origin query with numeric args replacement
-SELECT $p0, a, b, c WHERE id = $p0 AND date < $p1 AND value IN ($p2)`,
+			yql: `-- origin query with positional args replacement
+SELECT a, b, c WHERE id = $p0 AND date < $p1 AND value IN ($p2)`,
 			params: []interface{}{
 				table.ValueParam("$p0", types.Int32Value(1)),
 				table.ValueParam("$p1", types.TimestampValueFromTime(now)),
@@ -224,47 +148,47 @@ SELECT 1`,
 SELECT 1`,
 		},
 		{
-			sql: "SELECT $1, $2",
+			sql: "SELECT ?, ?",
 			args: []interface{}{
 				1,
 			},
 			err: ErrInconsistentArgs,
 		},
 		{
-			sql: "SELECT $1, $2 -- some comment with $3",
+			sql: "SELECT ?, ? -- some comment with ?",
 			args: []interface{}{
 				100,
 				200,
 			},
-			yql: `-- origin query with numeric args replacement
-SELECT $p0, $p1 -- some comment with $3`,
+			yql: `-- origin query with positional args replacement
+SELECT $p0, $p1 -- some comment with ?`,
 			params: []interface{}{
 				table.ValueParam("$p0", types.Int32Value(100)),
 				table.ValueParam("$p1", types.Int32Value(200)),
 			},
 		},
 		{
-			sql: "SELECT $1 /* some comment with $3 */, $2",
+			sql: "SELECT ? /* some comment with ? */, ?",
 			args: []interface{}{
 				100,
 				200,
 			},
-			yql: `-- origin query with numeric args replacement
-SELECT $p0 /* some comment with $3 */, $p1`,
+			yql: `-- origin query with positional args replacement
+SELECT $p0 /* some comment with ? */, $p1`,
 			params: []interface{}{
 				table.ValueParam("$p0", types.Int32Value(100)),
 				table.ValueParam("$p1", types.Int32Value(200)),
 			},
 		},
 		{
-			sql: "SELECT $1, $2 -- some comment with $3",
+			sql: "SELECT ?, ? -- some comment with ?",
 			args: []interface{}{
 				100,
 			},
 			err: ErrInconsistentArgs,
 		},
 		{
-			sql: "SELECT $1, $2, $3",
+			sql: "SELECT ?, ?, ?",
 			args: []interface{}{
 				100,
 				200,
@@ -273,26 +197,26 @@ SELECT $p0 /* some comment with $3 */, $p1`,
 		},
 		{
 			sql: `
-SELECT $1 /* some comment with $3 */, $2`,
+SELECT ? /* some comment with ? */, ?`,
 			args: []interface{}{
 				100,
 				200,
 			},
-			yql: `-- origin query with numeric args replacement
+			yql: `-- origin query with positional args replacement
 
-SELECT $p0 /* some comment with $3 */, $p1`,
+SELECT $p0 /* some comment with ? */, $p1`,
 			params: []interface{}{
 				table.ValueParam("$p0", types.Int32Value(100)),
 				table.ValueParam("$p1", types.Int32Value(200)),
 			},
 		},
 		{
-			sql: "SELECT $1, $2",
+			sql: "SELECT ?, ?",
 			args: []interface{}{
 				100,
 				200,
 			},
-			yql: `-- origin query with numeric args replacement
+			yql: `-- origin query with positional args replacement
 SELECT $p0, $p1`,
 			params: []interface{}{
 				table.ValueParam("$p0", types.Int32Value(100)),
@@ -300,12 +224,12 @@ SELECT $p0, $p1`,
 			},
 		},
 		{
-			sql: "SELECT $1, $2",
+			sql: "SELECT ?, ?",
 			args: []interface{}{
 				100,
 				200,
 			},
-			yql: `-- origin query with numeric args replacement
+			yql: `-- origin query with positional args replacement
 SELECT $p0, $p1`,
 			params: []interface{}{
 				table.ValueParam("$p0", types.Int32Value(100)),
