+++ internal/bind/sql_lexer.go
@@ -1,4 +1,4 @@
-package query
+package bind
 
 import "unicode/utf8"
