+++ sugar/params_go1.18_test.go
@@ -12,27 +12,30 @@ import (
 
 	"github.com/stretchr/testify/require"
 
+	"github.com/ydb-platform/ydb-go-sdk/v3/internal/bind"
 	"github.com/ydb-platform/ydb-go-sdk/v3/table"
 	"github.com/ydb-platform/ydb-go-sdk/v3/table/types"
+	"github.com/ydb-platform/ydb-go-sdk/v3/testutil"
 )
 
 func TestGenerateDeclareSection_ParameterOption(t *testing.T) {
-	splitDeclares := func(declaresSection string) (declares []string) {
-		for _, s := range strings.Split(declaresSection, ";") {
+	b := testutil.QueryBind(bind.AutoDeclare{})
+	getDeclares := func(declaresSection string) (declares []string) {
+		for _, s := range strings.Split(declaresSection, "\n") {
 			s = strings.TrimSpace(s)
-			if s != "" {
-				declares = append(declares, s)
+			if s != "" && !strings.HasPrefix(s, "--") {
+				declares = append(declares, strings.TrimRight(s, ";"))
 			}
 		}
 		sort.Strings(declares)
 		return declares
 	}
 	for _, tt := range []struct {
-		params  []table.ParameterOption
-		declare string
+		params   []interface{}
+		declares []string
 	}{
 		{
-			params: []table.ParameterOption{
+			params: []interface{}{
 				table.ValueParam(
 					"$values",
 					types.ListValue(
@@ -44,34 +47,34 @@ func TestGenerateDeclareSection_ParameterOption(t *testing.T) {
 					),
 				),
 			},
-			declare: `
-				DECLARE $values AS List<Uint64>;
-			`,
+			declares: []string{
+				"DECLARE $values AS List<Uint64>",
+			},
 		},
 		{
-			params: []table.ParameterOption{
+			params: []interface{}{
 				table.ValueParam(
 					"$delta",
 					types.IntervalValueFromDuration(time.Hour),
 				),
 			},
-			declare: `
-				DECLARE $delta AS Interval;
-			`,
+			declares: []string{
+				"DECLARE $delta AS Interval",
+			},
 		},
 		{
-			params: []table.ParameterOption{
+			params: []interface{}{
 				table.ValueParam(
 					"$ts",
 					types.TimestampValueFromTime(time.Now()),
 				),
 			},
-			declare: `
-				DECLARE $ts AS Timestamp;
-			`,
+			declares: []string{
+				"DECLARE $ts AS Timestamp",
+			},
 		},
 		{
-			params: []table.ParameterOption{
+			params: []interface{}{
 				table.ValueParam(
 					"$a",
 					types.BoolValue(true),
@@ -85,14 +88,14 @@ func TestGenerateDeclareSection_ParameterOption(t *testing.T) {
 					types.OptionalValue(types.TextValue("test")),
 				),
 			},
-			declare: `
-				DECLARE $a AS Bool;
-				DECLARE $b AS Int64; 
-				DECLARE $c AS Optional<Utf8>;
-			`,
+			declares: []string{
+				"DECLARE $a AS Bool",
+				"DECLARE $b AS Int64",
+				"DECLARE $c AS Optional<Utf8>",
+			},
 		},
 		{
-			params: []table.ParameterOption{
+			params: []interface{}{
 				table.ValueParam(
 					"$a",
 					types.BoolValue(true),
@@ -106,52 +109,39 @@ func TestGenerateDeclareSection_ParameterOption(t *testing.T) {
 					types.OptionalValue(types.TextValue("test")),
 				),
 			},
-			declare: `
-				DECLARE $a AS Bool;
-				DECLARE $b AS Int64; 
-				DECLARE $c AS Optional<Utf8>;
-			`,
+			declares: []string{
+				"DECLARE $a AS Bool",
+				"DECLARE $b AS Int64",
+				"DECLARE $c AS Optional<Utf8>",
+			},
 		},
 	} {
 		t.Run("", func(t *testing.T) {
-			declares, err := GenerateDeclareSection(tt.params)
+			yql, _, err := b.RewriteQuery("", tt.params...)
 			require.NoError(t, err)
-			got := splitDeclares(declares)
-			want := splitDeclares(tt.declare)
-			if len(got) != len(want) {
-				t.Errorf("len(got) = %v, len(want) = %v", len(got), len(want))
-			} else {
-				for i := range got {
-					if strings.TrimSpace(got[i]) != strings.TrimSpace(want[i]) {
-						t.Errorf(
-							"unexpected generation of declare section:\n%v\n\nwant:\n%v",
-							strings.Join(got, ";\n"),
-							strings.Join(want, ";\n"),
-						)
-					}
-				}
-			}
+			require.Equal(t, tt.declares, getDeclares(yql))
 		})
 	}
 }
 
 func TestGenerateDeclareSection_NamedArg(t *testing.T) {
-	splitDeclares := func(declaresSection string) (declares []string) {
-		for _, s := range strings.Split(declaresSection, ";") {
+	b := testutil.QueryBind(bind.AutoDeclare{})
+	getDeclares := func(declaresSection string) (declares []string) {
+		for _, s := range strings.Split(declaresSection, "\n") {
 			s = strings.TrimSpace(s)
-			if s != "" {
-				declares = append(declares, s)
+			if s != "" && !strings.HasPrefix(s, "--") {
+				declares = append(declares, strings.TrimRight(s, ";"))
 			}
 		}
 		sort.Strings(declares)
 		return declares
 	}
 	for _, tt := range []struct {
-		params  []sql.NamedArg
-		declare string
+		params   []interface{}
+		declares []string
 	}{
 		{
-			params: []sql.NamedArg{
+			params: []interface{}{
 				sql.Named(
 					"values",
 					types.ListValue(
@@ -163,34 +153,34 @@ func TestGenerateDeclareSection_NamedArg(t *testing.T) {
 					),
 				),
 			},
-			declare: `
-				DECLARE $values AS List<Uint64>;
-			`,
+			declares: []string{
+				"DECLARE $values AS List<Uint64>",
+			},
 		},
 		{
-			params: []sql.NamedArg{
+			params: []interface{}{
 				sql.Named(
 					"delta",
 					types.IntervalValueFromDuration(time.Hour),
 				),
 			},
-			declare: `
-				DECLARE $delta AS Interval;
-			`,
+			declares: []string{
+				"DECLARE $delta AS Interval",
+			},
 		},
 		{
-			params: []sql.NamedArg{
+			params: []interface{}{
 				sql.Named(
 					"ts",
 					types.TimestampValueFromTime(time.Now()),
 				),
 			},
-			declare: `
-				DECLARE $ts AS Timestamp;
-			`,
+			declares: []string{
+				"DECLARE $ts AS Timestamp",
+			},
 		},
 		{
-			params: []sql.NamedArg{
+			params: []interface{}{
 				sql.Named(
 					"a",
 					types.BoolValue(true),
@@ -204,14 +194,14 @@ func TestGenerateDeclareSection_NamedArg(t *testing.T) {
 					types.OptionalValue(types.TextValue("test")),
 				),
 			},
-			declare: `
-				DECLARE $a AS Bool;
-				DECLARE $b AS Int64; 
-				DECLARE $c AS Optional<Utf8>;
-			`,
+			declares: []string{
+				"DECLARE $a AS Bool",
+				"DECLARE $b AS Int64",
+				"DECLARE $c AS Optional<Utf8>",
+			},
 		},
 		{
-			params: []sql.NamedArg{
+			params: []interface{}{
 				sql.Named(
 					"a",
 					types.BoolValue(true),
@@ -225,71 +215,58 @@ func TestGenerateDeclareSection_NamedArg(t *testing.T) {
 					types.OptionalValue(types.TextValue("test")),
 				),
 			},
-			declare: `
-				DECLARE $a AS Bool;
-				DECLARE $b AS Int64; 
-				DECLARE $c AS Optional<Utf8>;
-			`,
+			declares: []string{
+				"DECLARE $a AS Bool",
+				"DECLARE $b AS Int64",
+				"DECLARE $c AS Optional<Utf8>",
+			},
 		},
 
 		{
-			params: []sql.NamedArg{
+			params: []interface{}{
 				sql.Named("delta", time.Hour),
 			},
-			declare: `
-				DECLARE $delta AS Interval;
-			`,
+			declares: []string{
+				"DECLARE $delta AS Interval",
+			},
 		},
 		{
-			params: []sql.NamedArg{
+			params: []interface{}{
 				sql.Named("ts", time.Now()),
 			},
-			declare: `
-				DECLARE $ts AS Timestamp;
-			`,
+			declares: []string{
+				"DECLARE $ts AS Timestamp",
+			},
 		},
 		{
-			params: []sql.NamedArg{
+			params: []interface{}{
 				sql.Named("$a", true),
 				sql.Named("$b", int64(123)),
 				sql.Named("$c", func(s string) *string { return &s }("test")),
 			},
-			declare: `
-				DECLARE $a AS Bool;
-				DECLARE $b AS Int64; 
-				DECLARE $c AS Optional<Utf8>;
-			`,
+			declares: []string{
+				"DECLARE $a AS Bool",
+				"DECLARE $b AS Int64",
+				"DECLARE $c AS Optional<Utf8>",
+			},
 		},
 		{
-			params: []sql.NamedArg{
+			params: []interface{}{
 				sql.Named("$a", func(b bool) *bool { return &b }(true)),
 				sql.Named("b", func(i int64) *int64 { return &i }(123)),
 				sql.Named("c", func(s string) *string { return &s }("test")),
 			},
-			declare: `
-				DECLARE $a AS Optional<Bool>;
-				DECLARE $b AS Optional<Int64>; 
-				DECLARE $c AS Optional<Utf8>;
-			`,
+			declares: []string{
+				"DECLARE $a AS Optional<Bool>",
+				"DECLARE $b AS Optional<Int64>",
+				"DECLARE $c AS Optional<Utf8>",
+			},
 		},
 	} {
 		t.Run("", func(t *testing.T) {
-			declares, err := GenerateDeclareSection(tt.params)
+			yql, _, err := b.RewriteQuery("", tt.params...)
 			require.NoError(t, err)
-			got := splitDeclares(declares)
-			want := splitDeclares(tt.declare)
-			if len(got) != len(want) {
-				t.Errorf("len(got) = %v, len(want) = %v", len(got), len(want))
-			} else {
-				for i := range got {
-					if strings.TrimSpace(got[i]) != strings.TrimSpace(want[i]) {
-						t.Errorf("unexpected generation of declare section:\n\n%v\n\nwant:\n%v",
-							strings.Join(got, ";\n"),
-							strings.Join(want, ";\n"),
-						)
-					}
-				}
-			}
+			require.Equal(t, tt.declares, getDeclares(yql))
 		})
 	}
 }
